refactor: extracted and cached all data loading
