+++ ncov19_dash/components/get_all_county_data.py
@@ -21,9 +21,13 @@ def get_all_county_data() -> pd.DataFrame:
     if response.status_code == 200:
         data = response.json()["message"]
         data = pd.DataFrame.from_records(data)
+        last_updated = data["last_update"][0]
     else:
         data = {"state_name": "john",
                 "county_name": "cena",
                 "confirmed": 0, "death": 0}
+        last_updated = "2100-01-01 00:00 EDT"
 
-    return data
+    return data, last_updated
+
+_, last_updated = get_all_county_data()#data["last_update"][0]
